refactor : 모델 테스트 클래스 별도 패키지로 이동
- 기존 모델을 테스트하던 테스트 클래스를 별도 패키지로 구분시킨다.
