+++ src/test/java/lotto/model/BuyPriceTest.java
@@ -1,6 +1,5 @@
-package lotto;
+package lotto.model;
 
-import lotto.model.BuyPrice;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
